requirements.txt
# core ml 9/25 
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# computer vission  9/25
opencv-python>=4.8.0
ultralytics>=8.0.0
mediapipe>=0.10.0

# pipelining 9/25
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# vizz
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# dev tools 9/25
jupyter>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0
pillow>=10.0.0

# tracking 9/25
filterpy>=1.4.5
scikit-image>=0.21.0

# utils 9/25
colorama>=0.4.6
rich>=13.5.0